chore: fillout sampling params (seed, n, best_of, min_p) (#3055)

+++ lib/llm/src/protocols/common.rs
@@ -274,14 +274,14 @@ pub const FREQUENCY_PENALTY_RANGE: (f32, f32) = (-1.0, 1.0);
 #[derive(Serialize, Deserialize, Debug, Clone, Default)]
 pub struct SamplingOptions {
     /// Number of output sequences to return for the given prompt
-    pub n: Option<i32>,
+    pub n: Option<u8>,
 
     /// Number of output sequences that are generated from the prompt.
     /// From these `best_of` sequences, the top `n` sequences are returned.
     /// `best_of` must be greater than or equal to `n`. This is treated as
     /// the beam width when `use_beam_search` is True. By default, `best_of`
     /// is set to `n`.
-    pub best_of: Option<i32>,
+    pub best_of: Option<u8>,
 
     /// Float that penalizes new tokens based on whether they
     /// appear in the generated text so far. Values > 0 encourage the model
